[Engine] Move Thread.h
